refactor reducers and actions and add react thunks
