--- src/actions/theme.ts
@@ -1,5 +1,0 @@
-import { ChangeThemeActionType } from '../types/default';
-
-export const selectTheme = (): ChangeThemeActionType => {
-  return { type: 'CHANGE_THEME' };
-};
